requirements.txt
matplotlib==3.2.2
numpy==1.18.5
pandas==1.0.5
Pillow==9.0.1
scikit_learn==1.0.2
tensorboardX==2.4.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.47.0
